// src/app/page/master/login-user-aplikasi/login-user-aplikasi.interface.ts
import {EventEmitter} from '@angular/core';

export interface LoginUserAplikasi {
	kdProfile?: string;
	kdAgama?: string;
	namaAgama?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
	getPage?: any;
	
}